Clean LLMFactory type structure
