+++ core/src/agents/workflows/orchestrator/nodes/inputNode.ts
@@ -1,7 +1,7 @@
-import { LLMFactory, LLMFactoryConfig } from '../../../../services/llm/factory.js';
+import { LLMFactory } from '../../../../services/llm/factory.js';
 import { createLogger } from '../../../../utils/logger.js';
 import { ApiConfig, OrchestratorStateType, Tools } from '../types.js';
-import { LLMConfiguration } from '../../../../services/llm/types.js';
+import { LLMConfiguration, LLMFactoryConfig } from '../../../../services/llm/types.js';
 import { ChatPromptTemplate } from '@langchain/core/prompts';
 import { attachLogger } from '../../../../api/server.js';
