+++ core/src/agents/workflows/orchestrator/nodes/messageSummaryNode.ts
@@ -1,8 +1,8 @@
 import { AIMessage } from '@langchain/core/messages';
 import { createLogger } from '../../../../utils/logger.js';
 import { ApiConfig, OrchestratorStateType } from '../types.js';
-import { LLMFactory, LLMFactoryConfig } from '../../../../services/llm/factory.js';
-import { LLMConfiguration } from '../../../../services/llm/types.js';
+import { LLMFactory } from '../../../../services/llm/factory.js';
+import { LLMConfiguration, LLMFactoryConfig } from '../../../../services/llm/types.js';
 import { PruningParameters } from '../types.js';
 import { ChatPromptTemplate } from '@langchain/core/prompts';
 import { attachLogger } from '../../../../api/server.js';
